--- java/placeholder.txt
@@ -1,1 +1,0 @@
-Delete me.
